move pwa code to pwa component
