+++ packages/admin-ui/src/pages/system/components/Notifications/Pwa.tsx
@@ -1,7 +1,70 @@
 import { api } from "api";
 import React, { useState, useEffect } from "react";
 import { Button } from "react-bootstrap";
-import { initializePushNotifications } from "registerServiceWorker.js";
+import { registerServiceWorker } from "registerServiceWorker.js";
+
+async function subscribeForPush(): Promise<void> {
+  try {
+    const swRegistration = await registerServiceWorker();
+    if (!swRegistration) {
+      console.error("Service Worker registration failed.");
+      return;
+    }
+    const pushManager = swRegistration.pushManager;
+    const existingSubscription = await pushManager.getSubscription();
+
+    if (existingSubscription) {
+      console.log("Existing subscription found:", existingSubscription);
+
+      // Ask backend if the subscription is still valid
+      const backendSubscription = await api.notificationsGetSubscription({ subscription: existingSubscription });
+
+      if (backendSubscription) {
+        console.log("Subscription is still valid. No need to resubscribe.");
+        return;
+      }
+
+      console.log("Subscription not found in the server. Unsubscribing...");
+      await existingSubscription.unsubscribe();
+    }
+
+    console.log("Requesting a new subscription...");
+    const newSubscription = await subscribeToPush(swRegistration);
+
+    if (newSubscription) {
+      console.log("New subscription obtained:", newSubscription);
+      await api.notificationsPostSubscription({ subscription: newSubscription });
+    }
+  } catch (error) {
+    console.error("Error subscribing for push notifications:", error);
+  }
+}
+
+async function subscribeToPush(swRegistration: ServiceWorkerRegistration): Promise<PushSubscription | undefined> {
+  try {
+    const VAPID_PUBLIC_KEY = await api.notificationsGetVapidPublicKey();
+    if (!VAPID_PUBLIC_KEY) {
+      console.error("Failed to retrieve VAPID public key.");
+      return;
+    }
+    const subscription = await swRegistration.pushManager.subscribe({
+      userVisibleOnly: true,
+      applicationServerKey: urlBase64ToUint8Array(VAPID_PUBLIC_KEY)
+    });
+    console.log("New Subscription:", subscription);
+    return subscription;
+  } catch (error) {
+    console.error("Push subscription failed:", error);
+    return undefined;
+  }
+}
+
+function urlBase64ToUint8Array(base64String: string): Uint8Array {
+  const padding = "=".repeat((4 - (base64String.length % 4)) % 4);
+  const base64 = (base64String + padding).replace(/-/g, "+").replace(/_/g, "/");
+  const rawData = atob(base64);
+  return new Uint8Array([...rawData].map((char) => char.charCodeAt(0)));
+}
 
 interface BeforeInstallPromptEvent extends Event {
   prompt: () => Promise<void>;
@@ -13,8 +76,18 @@ export default function Pwa() {
   const [deferredPrompt, setDeferredPrompt] = useState<BeforeInstallPromptEvent | null>(null);
   const [isPwaInstalled, setIsPwaInstalled] = useState(false);
 
+  // Check if permissions are granted
   useEffect(() => {
-    // Check if PWA is already installed
+    const checkNotificationsPermission = async () => {
+      const permission = await Notification.requestPermission();
+      setNotificationsGranted(permission);
+    };
+
+    checkNotificationsPermission();
+  }, []);
+
+  // Check if PWA is already installed
+  useEffect(() => {
     const checkPwaInstalled = () => {
       setIsPwaInstalled(window.matchMedia("(display-mode: standalone)").matches);
     };
@@ -52,7 +125,7 @@ export default function Pwa() {
     try {
       const permissionResult = await Notification.requestPermission();
       if (permissionResult === "granted") {
-        await initializePushNotifications();
+        await subscribeForPush();
         await api.notificationsPostNewNotification({
           title: "Notifications enabled",
           body: "You will now receive notifications from dappnode."
@@ -77,11 +150,10 @@ export default function Pwa() {
 
   return (
     <>
-      {!isPwaInstalled && deferredPrompt && (
-        <Button type="button" onClick={handleInstall} variant="dappnode">
-          Install PWA
-        </Button>
-      )}
+      <Button type="button" onClick={handleInstall} disabled={isPwaInstalled && !deferredPrompt} variant="dappnode">
+        Install PWA
+      </Button>
+
       <Button
         type="button"
         onClick={handleGrantNotifications}
